SCRIPT/requirements.txt
Flask>=2.0
openai>=2.0
python-dotenv
markdown
